requirements.txt
Flask==3.0.3
mongoengine==0.28.2
python-dotenv==1.0.1
gunicorn==22.0.0
